[step3] 파일 분리
